bootstrap.py: Remove --symlink option
A key design philosophy for DSI is that all configuration and
all log and result files reside in the same directory, known
ask the work directory. Being able to symlink to outside that
directory breaks this rule. A consequence could be that a tar file
capturing a specific test run, would contain just empty symlinks,
and the actual configuration used in the test is lost.