Can now run ycsb.short on CrateDB!
Includes a major sweep to keep cluster_setup.yml as generic
as possible. cluster_setup.meta.product_name is the key
that leads to defaults for each product. In an ironic twist,
the default is actually mongodb, even if MongoDB doesn't
use the generic cluster_setup yet. (But it almost could...)

Changing the other ycsb test_control.yml files is left to a
separate commit for simplicity. Can be done lazily as needed too.